revert: 画面名を汎用的な管理画面名に戻す
- ナビゲーション: 川崎汽船特化 → 汎用管理画面名に変更
- URL構成: 汎用的なパス (/customers, /products, /orders等) に復元
- 画面タイトル: 業務特化名 → シンプルな管理画面名に変更
- UIデザイン・機能は川崎汽船向けのままで維持

画面構成:
- 顧客管理: 顧客情報・取引履歴
- 商品管理: 在庫状況・価格管理
- 注文管理: 注文処理・ステータス管理
- 取引先管理: パートナー評価・契約管理
- 仕入先管理: 品質評価・認証管理
- 請求書管理: 支払い追跡・期限管理
- 在庫管理: 在庫レベル・発注管理
